fix: cursor jumps when editing definition properties
When definition properties are edited it fires `Modeling#updateProperties`. This method fires command `element.updateProperties` via commandStack to update element, same time `DrdUpdater` is subscribed to this command and do `DefinitionPropertiesView#update` when it executed.

When `DefinitionPropertiesView#update` is called it updates textContent of corresponding elements unconditionally which leads to cursor jump. In order to workaround issue with jumping cursor I've added condition that checks if element currently in focus and if so it prevent update of element content

Fixes #951
Fixes #274